Added Android support
I added Android support and a crossplatform base class that's extended when needed. This update was tested on my Android device and worked seamlessly with my Windows machine. In the next few updates, I will clean up the code, and add some automating features that will list all ip addresses without eating your ram, and listen for connections on the given port. I will also add a consent based system, where you are asked if you want to say copy a message given from a certain device, to avoid text-based hijacking (text-jacking?). In the next updates I will add a file transmission feature (hence the need for cleaning up the code), and plan on forcing the SDK to run all of its methods in the background to avoid delays