0.0.2: Metadata and other cool stuff
I added metadata support and file sharing support too. If I see it need be, I will make it so that the file downloads to a specific directory, and not where the server wants the file to be stored, so I don't know. I guess it depends on if this feature is used against the system. Automating and listing all clients on a wifi network will soon be available with nmap support, or by counting a loop and checking the address. If you have an idea, email me.